backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
coreapi==2.3.3
Pillow==10.1.0
